Misc polish, LivepeerAPIResourceWowza gets its own file
