+++ src/org/livepeer/LivepeerWowza/LivepeerAPI.java
@@ -2,10 +2,13 @@ package org.livepeer.LivepeerWowza;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.wowza.wms.logging.WMSLogger;
 import com.wowza.wms.rest.ShortObject;
+import com.wowza.wms.rest.vhosts.applications.streamfiles.StreamFileAppConfig;
+import com.wowza.wms.rest.vhosts.applications.streamfiles.StreamFilesAppConfig;
 import com.wowza.wms.rest.vhosts.applications.transcoder.TranscoderAppConfig;
 import com.wowza.wms.rest.vhosts.applications.transcoder.TranscoderTemplateAppConfig;
 import org.apache.http.HttpResponse;
@@ -117,8 +120,7 @@ public class LivepeerAPI {
    */
   public List<LivepeerAPIResourceBroadcaster> getBroadcasters() throws IOException {
     HttpResponse response = _get("/broadcaster");
-    TypeReference typeRef = new TypeReference<List<LivepeerAPIResourceBroadcaster>>() {
-    };
+    TypeReference typeRef = new TypeReference<List<LivepeerAPIResourceBroadcaster>>() {};
     List<LivepeerAPIResourceBroadcaster> list = mapper.readValue(response.getEntity().getContent(), typeRef);
     return list;
   }
@@ -141,63 +143,5 @@ public class LivepeerAPI {
       address = _address;
     }
 
-  }
-
-  @JsonAutoDetect(fieldVisibility = Visibility.ANY)
-  public static class LivepeerAPIResourceWowza {
-    private String id;
-    private TranscoderAppConfig transcoderAppConfig;
-    private Map<String, TranscoderTemplateAppConfig> transcoderTemplateAppConfig;
-
-    /**
-     * Create an empty API Resource. Mostly used by the JSON serializer when GETing objects
-     */
-    public LivepeerAPIResourceWowza() {
-
-    }
-
-    /**
-     * Create a Wowza resource from internal configuration, suitable for POSTing
-     *
-     * @param vhost
-     * @param application
-     */
-    public LivepeerAPIResourceWowza(String vhost, String application) {
-      id = null;
-      transcoderAppConfig = new TranscoderAppConfig(vhost, application);
-      transcoderAppConfig.loadObject();
-      transcoderTemplateAppConfig = new HashMap<>();
-      for (ShortObject t : transcoderAppConfig.getTemplates().getTemplates()) {
-        TranscoderTemplateAppConfig ttac = new TranscoderTemplateAppConfig(vhost, application, t.getId());
-        ttac.loadObject();
-        transcoderTemplateAppConfig.put(t.getId(), ttac);
-      }
-      ObjectMapper mapper = new ObjectMapper();
-
-    }
-
-    public String getId() {
-      return id;
-    }
-
-    public void setId(String _id) {
-      id = _id;
-    }
-
-    public TranscoderAppConfig getTranscoderAppConfig() {
-      return transcoderAppConfig;
-    }
-
-    public void setTranscoderAppConfig(TranscoderAppConfig tac) {
-      transcoderAppConfig = tac;
-    }
-
-    public Map<String, TranscoderTemplateAppConfig> getTranscoderTemplateAppConfig() {
-      return transcoderTemplateAppConfig;
-    }
-
-    public void setTranscoderTemplateAppConfig(Map<String, TranscoderTemplateAppConfig> ttac) {
-      transcoderTemplateAppConfig = ttac;
-    }
   }
 }
